added posts to data/db.json

+++ react-client/src/components/PostPage.jsx
@@ -9,13 +9,15 @@ const PostPage = (props) => {
   const post = posts.find(post => post.id.toString() === id)
 
   const handleDelete = (id) => {
-    api.delete(`/posts/${id}`)
+    if (window.confirm(`Are you sure you want to delete '${post.title}'? This action cannot be undone!`)) {
+      api.delete(`/posts/${id}`)
       .then(() => {
         const updatedPostList = posts.filter(post => post.id !== id)
         setPosts(updatedPostList)
         navigate('/')
       })
       .catch(err => console.log(err))
+    }
   }
 
   return (
